requirements.txt
python-dotenv==0.19.2
pika==1.2.0
SQLAlchemy==1.4.32
